+++ src/main/java/teamroots/embers/mixin/embersified/ModEventsMixin.java
@@ -1,21 +1,31 @@
 package teamroots.embers.mixin.embersified;
 
 import net.minecraft.block.Block;
+import net.minecraft.item.Item;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import p455w0rd.embersified.init.ModEvents;
 import teamroots.embers.register.BlockRegister;
+import teamroots.embers.register.ItemRegister;
 
 @Mixin(value = ModEvents.class, remap = false)
 public class ModEventsMixin {
     @Redirect(method = "onBlockRegistryReady", at = @At(value = "FIELD", target = "Lteamroots/embers/RegistryManager;ember_emitter:Lnet/minecraft/block/Block;"))
     private static void ember_emitter(Block value) {
+        BlockRegister.EMBER_EMITTER = value;
         BlockRegister.INSTANCE.add(value);
     }
 
     @Redirect(method = "onBlockRegistryReady", at = @At(value = "FIELD", target = "Lteamroots/embers/RegistryManager;ember_receiver:Lnet/minecraft/block/Block;"))
     private static void ember_receiver(Block value) {
+        BlockRegister.EMBER_RECEIVER = value;
         BlockRegister.INSTANCE.add(value);
     }
+
+    @Redirect(method = "onRecipeRegistryReady", at = @At(value = "FIELD", target = "Lteamroots/embers/RegistryManager;plate_caminite:Lnet/minecraft/item/Item;"))
+    private Item plate_caminite() {
+        return ItemRegister.PLATE_CAMINITE;
+    }
+
 }