+++ src/main/java/teamroots/embers/mixin/embersifiedextended/TileEjectorMixin.java
@@ -1,17 +1,18 @@
-package teamroots.embers.mixin.embersified;
+package teamroots.embers.mixin.embersifiedextended;
 
+import embersified.blocks.tiles.TileEjector;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.*;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import p455w0rd.embersified.blocks.tiles.TileEmitter;
 import teamroots.embers.config.ConfigMachine;
 
-@Mixin(value = TileEmitter.class, remap = false)
-public class TileEmitterMixin {
-
+@Mixin(value = TileEjector.class, remap = false)
+public class TileEjectorMixin {
     @Final
     @Mutable
     @Shadow
@@ -24,13 +25,12 @@ public class TileEmitterMixin {
 
     @ModifyArg(method = "<init>", at = @At(value = "INVOKE", target = "Lteamroots/embers/api/power/IEmberCapability;setEmberCapacity(D)V"), index = 0)
     private double modifyEmberCapacity(double capacity) {
-        return ConfigMachine.EMITTER.capacity;
+        return ConfigMachine.EJECTOR.capacity;
     }
 
     @Inject(method = "update", at = @At(value = "HEAD"))
     private void update(CallbackInfo info) {
-        TRANSFER_RATE = ConfigMachine.EMITTER.transfer_rate;
-        PULL_RATE = ConfigMachine.EMITTER.pull_rate;
+        TRANSFER_RATE = ConfigMachine.EJECTOR.transfer_rate;
+        PULL_RATE = ConfigMachine.EJECTOR.pull_rate;
     }
-
 }