+++ src/main/java/teamroots/embers/mixin/embersifiedextended/TileEmitterMixin.java
@@ -1,17 +1,18 @@
-package teamroots.embers.mixin.embersified;
+package teamroots.embers.mixin.embersifiedextended;
 
+import embersified.blocks.tiles.TileEmitter;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.*;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import p455w0rd.embersified.blocks.tiles.TileEmitter;
 import teamroots.embers.config.ConfigMachine;
 
 @Mixin(value = TileEmitter.class, remap = false)
 public class TileEmitterMixin {
-
     @Final
     @Mutable
     @Shadow
@@ -32,5 +33,4 @@ public class TileEmitterMixin {
         TRANSFER_RATE = ConfigMachine.EMITTER.transfer_rate;
         PULL_RATE = ConfigMachine.EMITTER.pull_rate;
     }
-
 }