+++ src/main/java/teamroots/embers/recipe/RecipeRegistry.java
@@ -22,6 +22,7 @@ import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.FluidUtil;
 import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
 import net.minecraftforge.fluids.capability.IFluidHandler;
+import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.items.ItemHandlerHelper;
@@ -437,7 +438,7 @@ public class RecipeRegistry {
                 "IPI",
                 'P', "plateIron",
                 'I', "ingotIron"}).setRegistryName(getRL("pipe")));
-        event.getRegistry().register(new ShapedOreRecipe(getRL("pump"), new ItemStack(BlockRegister.PIPE, 1), true, new Object[]{
+        event.getRegistry().register(new ShapedOreRecipe(getRL("pump"), new ItemStack(BlockRegister.PUMP, 1), true, new Object[]{
                 " R ",
                 "PBP",
                 " R ",
@@ -453,19 +454,21 @@ public class RecipeRegistry {
                 'F', Blocks.FURNACE,
                 'I', "ingotIron",
                 'C', "ingotCopper"}).setRegistryName(getRL("block_furnace")));
-        event.getRegistry().register(new ShapedOreRecipe(getRL("ember_receiver"), new ItemStack(BlockRegister.EMBER_RECEIVER, 4), true, new Object[]{
-                "I I",
-                "CPC",
-                'I', "ingotIron",
-                'C', "ingotCopper",
-                'P', ItemRegister.PLATE_CAMINITE}).setRegistryName(getRL("ember_receiver")));
-        event.getRegistry().register(new ShapedOreRecipe(getRL("ember_emitter"), new ItemStack(BlockRegister.EMBER_EMITTER, 4), true, new Object[]{
-                " C ",
-                " C ",
-                "IPI",
-                'I', "ingotIron",
-                'C', "ingotCopper",
-                'P', ItemRegister.PLATE_CAMINITE}).setRegistryName(getRL("ember_emitter")));
+        if (!Loader.isModLoaded("embersifiedextended") && !Loader.isModLoaded("embersified")) {
+            event.getRegistry().register(new ShapedOreRecipe(getRL("ember_receiver"), new ItemStack(BlockRegister.EMBER_RECEIVER, 4), true, new Object[]{
+                    "I I",
+                    "CPC",
+                    'I', "ingotIron",
+                    'C', "ingotCopper",
+                    'P', ItemRegister.PLATE_CAMINITE}).setRegistryName(getRL("ember_receiver")));
+            event.getRegistry().register(new ShapedOreRecipe(getRL("ember_emitter"), new ItemStack(BlockRegister.EMBER_EMITTER, 4), true, new Object[]{
+                    " C ",
+                    " C ",
+                    "IPI",
+                    'I', "ingotIron",
+                    'C', "ingotCopper",
+                    'P', ItemRegister.PLATE_CAMINITE}).setRegistryName(getRL("ember_emitter")));
+        }
         event.getRegistry().register(new ShapedOreRecipe(getRL("copper_cell"), new ItemStack(BlockRegister.COPPER_CELL, 1), true, new Object[]{
                 "BIB",
                 "ICI",
@@ -648,15 +651,17 @@ public class RecipeRegistry {
                 'S', ItemRegister.CRYSTAL_EMBER,
                 'C', "plateCopper",
                 'G', "blockGlass"}).setRegistryName(getRL("ember_cartridge")));
-        event.getRegistry().register(new ShapedOreRecipe(getRL("charger"), new ItemStack(BlockRegister.CHARGER, 1), true, new Object[]{
-                " X ",
-                "DCD",
-                "IPI",
-                'D', "ingotDawnstone",
-                'P', "plateCopper",
-                'C', "ingotCopper",
-                'I', "ingotIron",
-                'X', "plateIron"}).setRegistryName(getRL("charger")));
+        if (!Loader.isModLoaded("embersifiedextended")) {
+            event.getRegistry().register(new ShapedOreRecipe(getRL("charger"), new ItemStack(BlockRegister.CHARGER, 1), true, new Object[]{
+                    " X ",
+                    "DCD",
+                    "IPI",
+                    'D', "ingotDawnstone",
+                    'P', "plateCopper",
+                    'C', "ingotCopper",
+                    'I', "ingotIron",
+                    'X', "plateIron"}).setRegistryName(getRL("charger")));
+        }
         event.getRegistry().register(new ShapedOreRecipe(getRL("axe_clockwork"), new ItemStack(ItemRegister.AXE_CLOCKWORK, 1), true, new Object[]{
                 "PCP",
                 "ISI",
@@ -933,20 +938,22 @@ public class RecipeRegistry {
                 'C', ItemRegister.CRYSTAL_EMBER,
                 'P', "plateDawnstone",
                 'D', "ingotDawnstone"}).setRegistryName(getRL("caster_orb")));
-        event.getRegistry().register(new ShapedOreRecipe(getRL("ember_pulser"), new ItemStack(BlockRegister.EMBER_PULSER, 1), true, new Object[]{
-                "D",
-                "E",
-                "I",
-                'E', BlockRegister.EMBER_EMITTER,
-                'I', "ingotIron",
-                'D', "plateDawnstone"}).setRegistryName(getRL("ember_pulser")));
-        event.getRegistry().register(new ShapedOreRecipe(getRL("ember_funnel"), new ItemStack(BlockRegister.EMBER_FUNNEL, 1), true, new Object[]{
-                "D D",
-                "CRC",
-                " D ",
-                'R', BlockRegister.EMBER_RECEIVER,
-                'C', "ingotCopper",
-                'D', "plateDawnstone"}).setRegistryName(getRL("ember_funnel")));
+        if (!Loader.isModLoaded("embersifiedextended")) {
+            event.getRegistry().register(new ShapedOreRecipe(getRL("ember_pulser"), new ItemStack(BlockRegister.EMBER_PULSER, 1), true, new Object[]{
+                    "D",
+                    "E",
+                    "I",
+                    'E', BlockRegister.EMBER_EMITTER,
+                    'I', "ingotIron",
+                    'D', "plateDawnstone"}).setRegistryName(getRL("ember_pulser")));
+            event.getRegistry().register(new ShapedOreRecipe(getRL("ember_funnel"), new ItemStack(BlockRegister.EMBER_FUNNEL, 1), true, new Object[]{
+                    "D D",
+                    "CRC",
+                    " D ",
+                    'R', BlockRegister.EMBER_RECEIVER,
+                    'C', "ingotCopper",
+                    'D', "plateDawnstone"}).setRegistryName(getRL("ember_funnel")));
+        }
         event.getRegistry().register(new ShapedOreRecipe(getRL("caminite_lever"), new ItemStack(BlockRegister.CAMINITE_LEVER, 4), true, new Object[]{
                 "S",
                 "P",