+++ src/main/java/teamroots/embers/register/BlockRegister.java
@@ -92,13 +92,13 @@ public class BlockRegister implements IRegistrable<Block> {
     public static final Block DAWNSTONE_ANVIL = new BlockDawnstoneAnvil(Material.ANVIL, "dawnstone_anvil", true).setHarvestProperties("pickaxe", 1).setIsFullCube(false).setIsOpaqueCube(false).setHardness(1.6f).setLightOpacity(0);
     public static final Block EMBER_ACTIVATOR = new BlockActivator(Material.ROCK, "ember_activator", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);
     public static final Block EMBER_BORE = new BlockEmberBore(Material.ROCK, "ember_bore", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);
-    public static final Block EMBER_EMITTER = new BlockEmberEmitter(Material.ROCK, "ember_emitter", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(0.6f);
+    public static Block EMBER_EMITTER = new BlockEmberEmitter(Material.ROCK, "ember_emitter", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(0.6f);
     public static final Block EMBER_FUNNEL = new BlockEmberFunnel(Material.IRON, "ember_funnel", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);
     public static final Block EMBER_GAUGE = new BlockEmberGauge(Material.ROCK, "ember_gauge", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);
     public static final Block EMBER_INJECTOR = new BlockEmberInjector(Material.ROCK, "ember_injector", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);
     public static final Block EMBER_PIPE = new BlockEmberPipe(Material.IRON, "ember_pipe", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);
     public static final Block EMBER_PULSER = new BlockEmberPulser(Material.ROCK, "ember_pulser", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);
-    public static final Block EMBER_RECEIVER = new BlockEmberReceiver(Material.ROCK, "ember_receiver", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(0.6f);
+    public static Block EMBER_RECEIVER = new BlockEmberReceiver(Material.ROCK, "ember_receiver", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(0.6f);
     public static final Block EMBER_RELAY = new BlockRelay(Material.ROCK, "ember_relay", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);
     public static final Block EMBER_SIPHON = new BlockEmberSiphon(Material.ROCK, "ember_siphon", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);
     public static final Block FIELD_CHART = new BlockFieldChart(Material.ROCK, "field_chart", true).setIsFullCube(false).setIsOpaqueCube(false).setHarvestProperties("pickaxe", 0).setHardness(1.6f);